requirements.txt
azure-ai-ml
azure-cosmos
cuda-python
einops
imageio
natsort
opencv-python
git+https://github.com/openai/CLIP.git
git+https://github.com/facebookresearch/segment-anything.git
pandas
prefetch_generator
pillow
python-dotenv
pytorch_lightning
regex
matplotlib
numpy
pytest
scipy
spacy
setuptools
scikit-learn
scikit-image
smop
torch
torchio
torchsummary
torchvision
tqdm